+++ src/App.tsx
@@ -2,13 +2,16 @@ import { ThemeProvider } from 'styled-components'
 import { defaultTheme } from './styles/themes/default'
 import { GlobalStyle } from './styles/global'
 import { Transaction } from './pages/transactions'
+import { TransactionsProvider } from './contexts/transactionscontext'
 
 export function App() {
   return (
     <ThemeProvider theme={defaultTheme}>
       <GlobalStyle />
 
-      <Transaction />
+      <TransactionsProvider>
+        <Transaction />
+      </TransactionsProvider>
     </ThemeProvider>
   )
 }
