+++ src/components/summary/index.tsx
@@ -4,8 +4,12 @@ import {
   CurrencyDollar
 } from '@phosphor-icons/react'
 import { SummaryCard, SummaryContainer } from './styles'
+import { useContext } from 'react'
+import { TransactionsContext } from '../../contexts/transactionscontext'
 
 export function Summary() {
+  const { transactions } = useContext(TransactionsContext)
+
   return (
     <SummaryContainer>
       <SummaryCard>
